separate current_price into new table for further crawl need

+++ src/main/java/com/gable/currencyapp/dto/app/ResponseCoinDto.java
@@ -19,7 +19,7 @@ public class ResponseCoinDto {
   private boolean isDataFresh = false;
 
   @JsonIgnore
-  private long marketCapRank;
+  private int marketCapRank;
 
   private String image;
